fix formatting of identical text (#262)
Since the text that was matched by a search was used to get the actual
TextIter objects, in a buffer like "|asdf| |asdf|" only the first was
formatted bold "twice" and the second stayed normal. We now use the
MatchInfo position as the total offset to determine the TextIter
objects.

Also the multiple remove_tag() calls were replaced by a single
remove_all_tags() in the given range.